feat: enable PD disagg for dynamo + vllm + kvbm
