+++ components/backends/vllm/launch/agg_kvbm.sh
@@ -8,5 +8,5 @@ trap 'echo Cleaning up...; kill 0' EXIT
 python -m dynamo.frontend --http-port=8000 &
 
 # run worker with KVBM enabled
-DYN_KVBM_CPU_CACHE_GB=4 \
-  python -m dynamo.vllm --model Qwen/Qwen3-0.6B --connector kvbm
+DYN_LOG=debug DYN_KVBM_CPU_CACHE_GB=20 \
+  python -m dynamo.vllm --model Qwen/Qwen3-0.6B --connector kvbm --enforce-eager
