+++ lib/bindings/python/rust/llm/block_manager/vllm/connector/leader/slot.rs
@@ -499,6 +499,12 @@ impl Slot for VllmConnectorSlot {
         }
 
         // we should have enough device blocks to cover the newly scheduled tokens
+        tracing::debug!(
+            "ziqif current_position: {}; num_scheduled_tokens: {}; device_blocks.len(): {}",
+            self.current_position,
+            num_scheduled_tokens,
+            self.device_blocks.len()
+        );
         let next_position = self.current_position + num_scheduled_tokens;
         assert!(
             next_position <= self.device_blocks.len() * self.block_size,
