+++ lib/bindings/python/src/dynamo/llm/vllm_integration/connector/__init__.py
@@ -2,5 +2,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from .dynamo_connector import DynamoConnector, DynamoConnectorMetadata
+from .pd_connector import PdConnector, PdConnectorMetadata
 
-__all__ = ["DynamoConnector", "DynamoConnectorMetadata"]
+__all__ = [
+    "DynamoConnector",
+    "DynamoConnectorMetadata",
+    "PdConnector",
+    "PdConnectorMetadata",
+]
