+++ lib/bindings/python/src/dynamo/llm/vllm_integration/connector/dynamo_connector.py
@@ -85,6 +85,11 @@ class DynamoConnector(KVConnectorBase_V1):
     ) -> tuple[bool, Optional[dict[str, Any]]]:
         return self._scheduler.request_finished(request, block_ids)
 
+    # TODO(ziqif): comment out for now since MultiConnector does not allow different layouts for different connectors
+    # @classmethod
+    # def get_required_kvcache_layout(cls, vllm_config: "VllmConfig") -> Optional[str]:
+    #     return "NHD"
+
     # Worker
 
     def register_kv_caches(self, kv_caches: dict[str, torch.Tensor]):
